consolidate changelogs
